% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{auth_rtweet}
\alias{auth_rtweet}
\title{Ensure that \code{rtweet} is authenticated}
\usage{
auth_rtweet(...)
}
\arguments{
\item{...}{arguments passed to \code{rtweet::create_token()}}
}
\description{
This will call \code{rtweet::create_token()}. If a token is being created, the
following environment variables are used/needed:
}
\details{
\itemize{
\item \code{rtweet_api_key}
\item \code{rtweet_api_secret_key}
\item \code{rtweet_access_token}
\item \code{rtweet_access_token_secret}
}
}
