% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{connect}
\alias{connect}
\title{Connect to the sqlite tracking DB}
\usage{
connect(db_path, clean = FALSE)
}
\arguments{
\item{db_path}{path to the db}

\item{clean}{logical, should the db be unlinked and re-initialized}
}
\value{
connection
}
\description{
Connect to the sqlite tracking DB
}
