% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif-formatter.R
\name{format_exif}
\alias{format_exif}
\title{Format EXIF for tweet}
\usage{
format_exif(exif_data)
}
\arguments{
\item{exif_data}{EXIF data from \code{exifr::read_exif()}}
}
\value{
text for a tweet
}
\description{
Takes the EXIF data (from \code{exifr::read_exif()}) and formats it into a tweet.
}
