% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes_or_or_grapes}
\alias{grapes_or_or_grapes}
\alias{\%||\%}
\title{is null pipe}
\usage{
a \%||\% b
}
\arguments{
\item{a}{first thing}

\item{b}{what to do if the first thing is \code{NULL}}
}
\value{
a, unless a is \code{NULL} then b
}
\description{
borrowed from rlang
}
\keyword{internal}
