% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{queue}
\alias{queue}
\title{Queue photos for tweeting}
\usage{
queue(photo, con, proc_dir = "processed")
}
\arguments{
\item{photo}{path to the photo}

\item{con}{the database to use}

\item{proc_dir}{directory to place the processed files}
}
\value{
\code{TRUE} silently, if successful
}
\description{
Queue photo(s) for tweeting. This inserts them into the db but does not
actually tweet them out (yet)
}
