% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toot.R
\name{toot_collapse}
\alias{toot_collapse}
\title{Collapse toot text}
\usage{
toot_collapse(text)
}
\arguments{
\item{text}{text to flatten}
}
\value{
the text flattened with new lines
}
\description{
Collapse toot text
}
