% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toot.R
\name{toot_photo}
\alias{toot_photo}
\title{Toot a photo out}
\usage{
toot_photo(photo_df, path = NULL, ...)
}
\arguments{
\item{photo_df}{a dataframe with information about the photo to be tooted}

\item{path}{the path to the \code{phototweetr} directory (the one above "orig"
typically)}

\item{...}{arguments passed to \code{rtoot::post_toot}}
}
\value{
the \code{photo_df}, updated with toot time, etc.
}
\description{
Toot a photo out
}
