% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{tweet_collapse}
\alias{tweet_collapse}
\title{Collapse tweet text}
\usage{
tweet_collapse(text)
}
\arguments{
\item{text}{text to flatten}
}
\value{
the text flattened with new lines
}
\description{
Collapse tweet text
}
