% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{tweet_photo}
\alias{tweet_photo}
\title{Tweet a photo out}
\usage{
tweet_photo(photo_df, path = NULL, ...)
}
\arguments{
\item{photo_df}{a dataframe with information about the photo to be tweeted}

\item{path}{the path to the \code{phototweetr} directory (the one above "orig"
typically)}

\item{...}{arguments passed to \code{rtweet::post_tweet}}
}
\value{
the \code{photo_df}, updated with tweet time, etc.
}
\description{
Tweet a photo out
}
