% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{update_one}
\alias{update_one}
\title{Update a photo in the tracking db}
\usage{
update_one(photo_df, con)
}
\arguments{
\item{photo_df}{dataframe with updated photo information (with a single row)}

\item{con}{a sqlite connection}
}
\description{
Update a photo in the tracking db
}
