% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{weighted_coin}
\alias{weighted_coin}
\title{A weighted coin}
\usage{
weighted_coin(positive = 1, negative = 14)
}
\arguments{
\item{positive}{the odds of \code{TRUE} (default: 1)}

\item{negative}{the offs of \code{FALSE} (default: 14)}
}
\value{
logical
}
\description{
Flips a coin that is weighted
}
